refactor: [#200] extract generate_environment_config to shared task
- Create generate_config.rs in src/testing/e2e/tasks/black_box/
- Function generates environment config with absolute SSH key paths
- Remove duplicate implementations from both E2E binaries
