+++ src/bin/e2e_provision_and_destroy_tests.rs
@@ -40,14 +40,13 @@
 
 use anyhow::Result;
 use clap::Parser;
-use std::path::PathBuf;
 use std::time::Instant;
 use torrust_dependency_installer::Dependency;
 use tracing::{error, info, warn};
 
 use torrust_tracker_deployer_lib::bootstrap::logging::{LogFormat, LogOutput, LoggingBuilder};
 use torrust_tracker_deployer_lib::testing::e2e::tasks::black_box::{
-    run_preflight_cleanup, verify_required_dependencies,
+    generate_environment_config, run_preflight_cleanup, verify_required_dependencies,
 };
 use torrust_tracker_deployer_lib::testing::e2e::ProcessRunner;
 
@@ -182,81 +181,6 @@ fn run_e2e_test_workflow(environment_name: &str, destroy: bool) -> Result<()> {
     }
 
     Ok(())
-}
-
-/// Generates the environment configuration file with absolute SSH key paths.
-///
-/// This function creates a temporary configuration file with absolute paths
-/// to the SSH keys, ensuring they work correctly regardless of the directory
-/// from which Ansible runs.
-///
-/// # Arguments
-///
-/// * `environment_name` - The name of the environment to create
-///
-/// # Returns
-///
-/// Returns the path to the generated configuration file.
-///
-/// # Errors
-///
-/// Returns an error if the configuration file cannot be created.
-fn generate_environment_config(environment_name: &str) -> Result<PathBuf> {
-    use std::fs;
-
-    // Get project root from current directory (cargo run runs from project root)
-    let project_root = std::env::current_dir()
-        .map_err(|e| anyhow::anyhow!("Failed to get current directory: {e}"))?;
-
-    // Build absolute paths to SSH keys
-    let private_key_path = project_root.join("fixtures/testing_rsa");
-    let public_key_path = project_root.join("fixtures/testing_rsa.pub");
-
-    // Verify SSH keys exist
-    if !private_key_path.exists() {
-        return Err(anyhow::anyhow!(
-            "SSH private key not found at: {}",
-            private_key_path.display()
-        ));
-    }
-    if !public_key_path.exists() {
-        return Err(anyhow::anyhow!(
-            "SSH public key not found at: {}",
-            public_key_path.display()
-        ));
-    }
-
-    // Create configuration JSON with absolute paths
-    let config = serde_json::json!({
-        "environment": {
-            "name": environment_name
-        },
-        "ssh_credentials": {
-            "private_key_path": private_key_path.to_string_lossy(),
-            "public_key_path": public_key_path.to_string_lossy()
-        }
-    });
-
-    // Write to envs directory
-    let config_path = project_root.join(format!("envs/{environment_name}.json"));
-
-    // Ensure parent directory exists
-    if let Some(parent) = config_path.parent() {
-        fs::create_dir_all(parent)
-            .map_err(|e| anyhow::anyhow!("Failed to create config directory: {e}"))?;
-    }
-
-    fs::write(&config_path, serde_json::to_string_pretty(&config)?)
-        .map_err(|e| anyhow::anyhow!("Failed to write config file: {e}"))?;
-
-    info!(
-        config_path = %config_path.display(),
-        private_key = %private_key_path.display(),
-        public_key = %public_key_path.display(),
-        "Generated environment configuration with absolute SSH key paths"
-    );
-
-    Ok(config_path)
 }
 
 /// Creates the environment from the configuration file.
